[tests] move some classes for tests into a testslib directory
