+++ tests/units/classes/Request/OGCRequestTest.php
@@ -2,8 +2,6 @@
 
 use PHPUnit\Framework\TestCase;
 
-require_once __DIR__.'/ClassesForTests.php';
-
 class OGCRequestTest extends TestCase
 {
     public function testParam()
@@ -13,7 +11,7 @@ class OGCRequestTest extends TestCase
             'service' => 'WMS',
             'empty' => ''
         );
-        $ogc = new OGCRequestForTest(new ProjectForOGC(), $params, null, new testContext());
+        $ogc = new OGCRequestForTests(new ProjectForOGCForTests(), $params, null, new ContextForTests());
         foreach ($params as $key => $value) {
             $this->assertEquals($value, $ogc->param($key));
         }
@@ -39,12 +37,12 @@ class OGCRequestTest extends TestCase
             'Lizmap_User' => '',
             'Lizmap_User_Groups' => ''
         );
-        $testContext = new testContext();
+        $testContext = new ContextForTests();
         $testContext->setResult($contextResult);
-        $ogc = new OGCRequestForTest(new ProjectForOGC(), $params, null, $testContext);
+        $ogc = new OGCRequestForTests(new ProjectForOGCForTests(), $params, null, $testContext);
         $parameters = $ogc->parameters();
         $this->assertEquals($expectedParameters, $parameters);
-        $proj = new ProjectForOGC();
+        $proj = new ProjectForOGCForTests();
         $proj->setRepo(new Lizmap\Project\Repository('test', array(), '', null, $testContext));
         $contextResult = array(
             'userIsConnected' => true,
@@ -60,7 +58,7 @@ class OGCRequestTest extends TestCase
             'Lizmap_User_Groups' => 'admins, users',
             'Lizmap_Override_Filter' => true
         );
-        $ogc = new OGCRequestForTest($proj, $params, null, $testContext);
+        $ogc = new OGCRequestForTests($proj, $params, null, $testContext);
         $parameters = $ogc->parameters();
         $this->assertEquals($expectedParameters, $parameters);
     }
@@ -71,18 +69,18 @@ class OGCRequestTest extends TestCase
             'service' => 'WMS',
             'request' => 'getcapabilities'
         );
-        $ogc = $this->getMockBuilder(OGCRequestForTest::class)
+        $ogc = $this->getMockBuilder(OGCRequestForTests::class)
             ->setMethods(['process_getcapabilities'])
-            ->setConstructorArgs([new ProjectForOGC(), $params, null, new testContext()])->getMock();
+            ->setConstructorArgs([new ProjectForOGCForTests(), $params, null, new ContextForTests()])->getMock();
         $ogc->expects($this->once())->method('process_getcapabilities');
         $ogc->process();
         $params = array(
             'service' => 'WMS',
             'request' => 'not existing method'
         );
-        $ogc = $this->getMockBuilder(OGCRequestForTest::class)
+        $ogc = $this->getMockBuilder(OGCRequestForTests::class)
             ->setMethods(['serviceException'])
-            ->setConstructorArgs([new ProjectForOGC(), $params, null, new testContext()])->getMock();
+            ->setConstructorArgs([new ProjectForOGCForTests(), $params, null, new ContextForTests()])->getMock();
         $ogc->expects($this->once())->method('serviceException')->with(501);
         $testMock = $ogc->process();
     }
