+++ tests/units/testslib/LayerWFSForTests.php
@@ -1,0 +1,9 @@
+<?php
+
+class LayerWFSForTests
+{
+    public function getSrid()
+    {
+        return 'SRID';
+    }
+}
