+++ tests/units/testslib/OGCRequestForTests.php
@@ -1,0 +1,7 @@
+<?php
+
+use Lizmap\Request\OGCRequest;
+
+class OGCRequestForTests extends OGCRequest
+{
+}
