+++ tests/units/testslib/ProjectForOGCForTests.php
@@ -1,0 +1,21 @@
+<?php
+
+
+class ProjectForOGCForTests extends ProjectForTests
+{
+    public $loginFilters;
+
+    public function getRelativeQgisPath()
+    {
+    }
+
+    public function getLoginFilters($layerName, $edition = false)
+    {
+        return $this->loginFilters;
+    }
+
+    public function setData($key, $value)
+    {
+        $this->data[$key] = $value;
+    }
+}
