+++ tests/units/testslib/QgisForm2ForTests.php
@@ -1,0 +1,10 @@
+<?php
+
+use Lizmap\Form;
+
+class QgisForm2ForTests extends Form\QgisForm
+{
+    public function getAttributesEditorForm()
+    {
+    }
+}
