+++ tests/units/testslib/WMTSRequestForTests.php
@@ -1,0 +1,12 @@
+<?php
+
+use Lizmap\Request\WMTSRequest;
+
+
+class WMTSRequestForTests extends WMTSRequest
+{
+    public function getCapabilitiesForTests()
+    {
+        return $this->process_getcapabilities();
+    }
+}
