+++ tests/units/testslib/jDbConnectionForTests.php
@@ -1,0 +1,14 @@
+<?php
+
+class jDbConnectionForTests
+{
+    public function encloseName($name)
+    {
+        return $name;
+    }
+
+    public function quote($name)
+    {
+        return '"'.$name.'"';
+    }
+}
