+++ tests/units/testslib/jTplForTests.php
@@ -1,0 +1,15 @@
+<?php
+
+
+class jTplForTests
+{
+    public function assign($arg1, $arg2)
+    {
+
+    }
+
+    public function fetch($arg1)
+    {
+
+    }
+}
